+++ src/components.d.ts
@@ -30,6 +30,7 @@ export namespace Components {
   interface Page7ReasonsWhyTypescriptHelpsYouKickAss {}
   interface PageAJavascriptImplementationOfJavaUtilMap {}
   interface PageAbout {}
+  interface PageAboutZsphereSmoothingBrushes {}
   interface PageAccessTheWasConsoleFromWebspherePortalServer {}
   interface PageAccessingTheAuthenticatedWebspherePortalUserFromASeparateWebApp {}
   interface PageApplyDifferentMaterialsToSubtoolsInZbrush {}
@@ -198,6 +199,12 @@ declare global {
   var HTMLPageAboutElement: {
     prototype: HTMLPageAboutElement;
     new (): HTMLPageAboutElement;
+  };
+
+  interface HTMLPageAboutZsphereSmoothingBrushesElement extends Components.PageAboutZsphereSmoothingBrushes, HTMLStencilElement {}
+  var HTMLPageAboutZsphereSmoothingBrushesElement: {
+    prototype: HTMLPageAboutZsphereSmoothingBrushesElement;
+    new (): HTMLPageAboutZsphereSmoothingBrushesElement;
   };
 
   interface HTMLPageAccessTheWasConsoleFromWebspherePortalServerElement extends Components.PageAccessTheWasConsoleFromWebspherePortalServer, HTMLStencilElement {}
@@ -805,6 +812,7 @@ declare global {
     'page-7-reasons-why-typescript-helps-you-kick-ass': HTMLPage7ReasonsWhyTypescriptHelpsYouKickAssElement;
     'page-a-javascript-implementation-of-java-util-map': HTMLPageAJavascriptImplementationOfJavaUtilMapElement;
     'page-about': HTMLPageAboutElement;
+    'page-about-zsphere-smoothing-brushes': HTMLPageAboutZsphereSmoothingBrushesElement;
     'page-access-the-was-console-from-websphere-portal-server': HTMLPageAccessTheWasConsoleFromWebspherePortalServerElement;
     'page-accessing-the-authenticated-websphere-portal-user-from-a-separate-web-app': HTMLPageAccessingTheAuthenticatedWebspherePortalUserFromASeparateWebAppElement;
     'page-apply-different-materials-to-subtools-in-zbrush': HTMLPageApplyDifferentMaterialsToSubtoolsInZbrushElement;
@@ -928,6 +936,7 @@ declare namespace LocalJSX {
   interface Page7ReasonsWhyTypescriptHelpsYouKickAss {}
   interface PageAJavascriptImplementationOfJavaUtilMap {}
   interface PageAbout {}
+  interface PageAboutZsphereSmoothingBrushes {}
   interface PageAccessTheWasConsoleFromWebspherePortalServer {}
   interface PageAccessingTheAuthenticatedWebspherePortalUserFromASeparateWebApp {}
   interface PageApplyDifferentMaterialsToSubtoolsInZbrush {}
@@ -1040,6 +1049,7 @@ declare namespace LocalJSX {
     'page-7-reasons-why-typescript-helps-you-kick-ass': Page7ReasonsWhyTypescriptHelpsYouKickAss;
     'page-a-javascript-implementation-of-java-util-map': PageAJavascriptImplementationOfJavaUtilMap;
     'page-about': PageAbout;
+    'page-about-zsphere-smoothing-brushes': PageAboutZsphereSmoothingBrushes;
     'page-access-the-was-console-from-websphere-portal-server': PageAccessTheWasConsoleFromWebspherePortalServer;
     'page-accessing-the-authenticated-websphere-portal-user-from-a-separate-web-app': PageAccessingTheAuthenticatedWebspherePortalUserFromASeparateWebApp;
     'page-apply-different-materials-to-subtools-in-zbrush': PageApplyDifferentMaterialsToSubtoolsInZbrush;
@@ -1159,6 +1169,7 @@ declare module "@stencil/core" {
       'page-7-reasons-why-typescript-helps-you-kick-ass': LocalJSX.Page7ReasonsWhyTypescriptHelpsYouKickAss & JSXBase.HTMLAttributes<HTMLPage7ReasonsWhyTypescriptHelpsYouKickAssElement>;
       'page-a-javascript-implementation-of-java-util-map': LocalJSX.PageAJavascriptImplementationOfJavaUtilMap & JSXBase.HTMLAttributes<HTMLPageAJavascriptImplementationOfJavaUtilMapElement>;
       'page-about': LocalJSX.PageAbout & JSXBase.HTMLAttributes<HTMLPageAboutElement>;
+      'page-about-zsphere-smoothing-brushes': LocalJSX.PageAboutZsphereSmoothingBrushes & JSXBase.HTMLAttributes<HTMLPageAboutZsphereSmoothingBrushesElement>;
       'page-access-the-was-console-from-websphere-portal-server': LocalJSX.PageAccessTheWasConsoleFromWebspherePortalServer & JSXBase.HTMLAttributes<HTMLPageAccessTheWasConsoleFromWebspherePortalServerElement>;
       'page-accessing-the-authenticated-websphere-portal-user-from-a-separate-web-app': LocalJSX.PageAccessingTheAuthenticatedWebspherePortalUserFromASeparateWebApp & JSXBase.HTMLAttributes<HTMLPageAccessingTheAuthenticatedWebspherePortalUserFromASeparateWebAppElement>;
       'page-apply-different-materials-to-subtools-in-zbrush': LocalJSX.PageApplyDifferentMaterialsToSubtoolsInZbrush & JSXBase.HTMLAttributes<HTMLPageApplyDifferentMaterialsToSubtoolsInZbrushElement>;